requirements.txt
aiohttp==2.3.10
asn1crypto==0.24.0
astroid==1.6.3
async-lru==0.1.0
async-timeout==2.0.1
attrdict==2.0.0
Babel==2.5.3
bitcoin==1.1.42
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
cytoolz==0.9.0.1
Django==2.0.4
django-phonenumber-field==2.0.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.8.2
eth-abi==1.0.0
eth-account==0.1.0a2
eth-bloom==1.0.0
eth-hash==0.1.2
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.0
eth-tester==0.1.0b23
eth-testrpc==1.3.3
eth-utils==1.0.3
ethereum==1.6.1
hexbytes==0.1.0
idna==2.6
idna-ssl==1.0.1
isort==4.3.4
json-rpc==1.10.8
lazy-object-proxy==1.3.1
lru-dict==1.1.6
mccabe==0.6.1
multidict==4.2.0
openassets==1.3
pbkdf2==1.3
phonenumberslite==8.9.4
py-ecc==1.4.2
py-evm==0.2.0a14
py-geth==2.0.1
py-solc==2.1.0
pycparser==2.18
pycryptodome==3.6.1
pyethash==0.1.27
pylint==1.8.4
pylint-django==0.11
pylint-plugin-utils==0.2.6
pysha3==1.0.2
python-bitcoinlib==0.2.1
pytz==2018.4
PyYAML==3.12
repoze.lru==0.7
requests==2.18.4
rlp==0.6.0
scrypt==0.8.6
secp256k1==0.13.2
semantic-version==2.6.0
six==1.11.0
toolz==0.9.0
trie==1.3.4
urllib3==1.22
web3==4.1.0
websockets==4.0.1
Werkzeug==0.14.1
wrapt==1.10.11
yarl==1.1.1
